A2/requirements.txt
Flask==3.0.3
mysql-connector-python
